Moving test utility files outside cpp test folder
