+++ kratos/tests/test_utilities/test_bar_element.cpp
@@ -19,7 +19,7 @@
 #include "includes/define.h"
 #include "includes/expect.h"
 #include "includes/variables.h"
-#include "tests/cpp_tests/auxiliar_files_for_cpp_unnitest/test_bar_element.h"
+#include "tests/test_utilities/test_bar_element.h"
 
 namespace Kratos::Testing
 {
