+++ kratos/tests/test_utilities/test_element.cpp
@@ -18,7 +18,7 @@
 #include "includes/define.h"
 #include "includes/expect.h"
 #include "includes/variables.h"
-#include "tests/cpp_tests/auxiliar_files_for_cpp_unnitest/test_element.h"
+#include "tests/test_utilities/test_element.h"
 
 namespace Kratos::Testing
 {
