+++ kratos/tests/test_utilities/test_included_parameters.json
@@ -1,0 +1,8 @@
+{
+    "level1":
+    {
+    "list_value":[ 3, "hi", false],
+    "tmp" : 5.0
+    },
+    "@include_json" : "testing_utilities/test_included_parameters_level2.json"
+}
